+++ docs/conf.py
@@ -11,7 +11,6 @@ sys.path.insert(0, os.path.abspath('..'))
 extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.doctest',
-    'sphinx_autodoc_typehints',
     'sphinx_paramlinks',
 ]